% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{crossValidation}
\alias{crossValidation}
\title{Cross Validation}
\usage{
crossValidation(data, k, ValidationFunction, nrRuns, ...)
}
\arguments{
\item{data}{a data.frame which includes follow up time, event, and covariates.}

\item{k}{a numeric value specifying k-fold cross validation.}

\item{nrRuns}{a numeric value indicating the number of runs of cross validation is to be performed.}

\item{...}{additional arguments to ValidationFunciton, see \link[survivalValidation]{IBS},
\link[survivalValidation]{calibration},\link[survivalValidation]{ROC}. Could also be a userspecified ValidationFunction}

\item{ValidationFunciton}{a function which validates a given model, inbuilt funcitons are IBS,ROC and Calibration.}
}
\value{
The function returns a large list of length nrRuns. Each element in the list contains all k elements from one run of cross validation.
}
\description{
Cross Validation
}
\examples{
add(1,2)
}

