% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_validation.R
\name{summarize_validation}
\alias{summarize_validation}
\title{Summarize Validation method
Computes aggregated value from one of the validation methods, \link[survivalValidation]{IBS},
\link[survivalValidation]{calibration},\link[survivalValidation]{ROC},\link[survivalValidation]{c_statistic}.}
\usage{
summarize_validation(cvList, validationMethod, simplify = FALSE)
}
\description{
Summarize Validation method
Computes aggregated value from one of the validation methods, \link[survivalValidation]{IBS},
\link[survivalValidation]{calibration},\link[survivalValidation]{ROC},\link[survivalValidation]{c_statistic}.
}

