% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBS.R
\name{IBS}
\alias{IBS}
\title{Integrated Brier Score}
\usage{
IBS(train, test, FittingFunction, covariates, ...)
}
\arguments{
\item{train}{a data.frame which includes follow up time (fu), event, and covariates. In a k-fold cross validation this is a sample proportion (k-1)/k of the original data set.}

\item{test}{a data.frame which includes follow up time (fu), event, and covariates. In a k-fold cross validation this is a sample proportion 1/k of the original data set.}

\item{FittingFunction}{a function which returns list of survival probabilities of individuals where each element in the list represent a specific time point.}

\item{covariates}{character vector specifying the names of covariates}
}
\description{
Integrated Brier Score
}
\examples{
IBS(train)
}

